% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{addColnames}
\alias{addColnames}
\title{Add Column Names}
\usage{
addColnames(matrix, colnames)
}
\arguments{
\item{colnames}{The new column names}

\item{matix}{The data that the new column names are being added to}
}
\value{
matrix with the same values but new col.name[s]
}
\description{
functionalizes colnames(matrix)<-newnames operation.
}
\examples{
addColnames(cbind(c(1,2,3),c(4,5,6)),c("low","high"))
}
\seealso{
\code{\link{addRownames} \link{addNames}}
}

