% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCCA.r
\name{addFasta}
\alias{addFasta}
\title{Add Fasta}
\usage{
addFasta(env, genome = BSgenome.Hsapiens.UCSC.hg19, width = 300)
}
\arguments{
\item{env}{A list containing the location of the peaks}

\item{genome}{the genome to which the locations apply}

\item{width}{the width of the peaks}
}
\value{
env with an attached fasta parameter
}
\description{
Appends a Biostrings Object to ENV containing the base pair infromation
for each peak. The input env list is reqired to have a bed value that
contains a start and an chr. The mean of the start and end values is
extended by the width and used to find the base pair infromation of
interest
}

