% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{addNames}
\alias{addNames}
\title{Add Names}
\usage{
addNames(matrix, colnames, rownames = colnames, list = NULL)
}
\arguments{
\item{matrix}{input matrix or list}

\item{colnames}{new colnames}

\item{rownames}{new rownmaes}

\item{list}{if not NULL matrix is a list}
}
\value{
a list with new names
}
\description{
Adds row and column names to a matrix
}
\examples{
addNames(cbind(c(1,2,3),c(4,5,6)),c("high","low"),c("a","b","c"))
addNames(list(1,2,3),c("a","b","c"),list=TRUE)
}
\seealso{
\code{\link{addColnames} \link{addRownames}}
}

