% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{addReg}
\alias{addReg}
\alias{addReg.default}
\title{AddReg Plot}
\usage{
addReg(x, tag, logic, ...)

\method{addReg}{default}()
}
\arguments{
\item{x}{The object with a $reg value to be added}

\item{tag}{The moniker of the new region to be added}

\item{logic}{A logical list of equal length to other regions in x$reg}
}
\description{
AddReg Plot
}

