% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{addRownames}
\alias{addRownames}
\title{Add Row Names}
\usage{
addRownames(matrix, rownames)
}
\arguments{
\item{matix}{The data that the new row names are being added to}

\item{colnames}{The new row names}
}
\value{
matrix with the same values but new row.name[s]
}
\description{
functionalizes rownames(matrix)<-newnames operation.
}
\examples{
addRownames(cbind(c(1,2,3),c(4,5,6)),c("a","b","c"))
}
\seealso{
\code{\link{addColnames} \link{addNames}}
}

