% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{applyPeakPartitions}
\alias{applyPeakPartitions}
\title{apply peak partitions}
\usage{
applyPeakPartitions(prc, inObj)
}
\arguments{
\item{prc}{the vectors}

\item{inObj}{list(list(which vector, function, ns) ...)}
}
\value{
the regions seperated
}
\description{
a wrapper for peak partitions that takes an inObj that contains all
of the information nessicary for the analysis
}

