% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.data.ccf.functions.r
\name{Bproj.summary.f}
\alias{Bproj.summary.f}
\title{Summarise a projection with multiple Monte Carlo realisations to make a dataframe with quantiles for plotting}
\usage{
Bproj.summary.f(PB, Bproj, PBproj, Eproj)
}
\arguments{
\item{PB}{the data and model fit coming from applying the PB model (PB.f)}

\item{Bproj}{the projected biomass for the climate and fishing scenario}

\item{PBproj}{the projected P/B ratio for the climate scenario}

\item{Eproj}{the projected climate scenario}
}
\description{
Summarise a projection with multiple Monte Carlo realisations to make a dataframe with quantiles for plotting
}
\examples{
proj.summary.f(PB,Bproj,PBproj,Eproj)
}
\keyword{confidence}
\keyword{interval}
\keyword{projection,}
\keyword{quantile,}
