% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccca.r
\name{ccca}
\alias{ccca}
\title{CCCA}
\usage{
ccca(dataSets, peakLists, categories)
}
\arguments{
\item{dataSets}{list of files containing the raw reads in bed format}

\item{peakLists}{list of files containing the peaks reads in MACS xls format}

\item{categories}{the moniker to be attached to each of the data set files}
}
\value{
a list of class ccca containing afs,udm, and prc
}
\description{
CCCA
}
\examples{
dataSets<-c("raw_sample1.bed","raw_sample2.bed","raw_sample3.bed")
peakList<-c("sample1.xls","sample2.xls","sample3.xls")
categories<-c("s1","s2","s3")
ccca<-ccca(dataSets,peakLists,categories)
## Find dimensions that seperate data sets of interest
plot(ccca,1,2)
plot(ccca,1,3)
ccca<-local({
  nm<-normalize(ccca$prc)
  ccca<-addReg(ccca,"s1",nm[,1]<(mean(nm[,1])-3*sd(nm[,1])))
  ccca<-addReg(ccca,"s2",nm[,1]>(mean(nm[,1])+3*sd(nm[,1])))
  ccca<-addReg(ccca,"s3",nm[,2]>(mean(nm[,2])+3*sd(nm[,2])))
  ccca<-addReg(ccca,"s1.me",ccca$reg[,"s1"] & !(ccca$reg[,"s2"] | ccca$reg[,"s3"]))
  ccca<-addReg(ccca,"s2.me",ccca$reg[,"s2"] & !(ccca$reg[,"s1"] | ccca$reg[,"s3"]))
  ccca<-addReg(ccca,"s3.me",ccca$reg[,"s3"] & !(ccca$reg[,"s2"] | ccca$reg[,"s1"]))
  ccca
})
ccca<-addFasta(ccca,genome)
s1Fasta<-ccca$fasta[ccca$reg[,"s1.me"]]
}

