% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{CCCA_prc}
\alias{CCCA_prc}
\title{CCCA prc}
\format{A CCCA Object}
\usage{
CCCA_prc
}
\value{
A CCCA Object for samples 1 - 3
}
\description{
Sample CCCA object generated from samples 1-3
}
\keyword{datasets}
