% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height.r
\name{chromGenomeSort}
\alias{chromGenomeSort}
\title{Chrom Genome Sort}
\usage{
chromGenomeSort(chromOrder)
}
\arguments{
\item{chromOrder}{A list of chromosomes to be found in the data}
}
\description{
Returns a function that reorders the peak data based on the chromosome order provided
}
\examples{
sortedChroms<-strsplit("chr1 chr2 chr3 chr4 chr5 chr6 chr7 chrX
chr8 chr9 chr10 chr11 chr12 chr13 chr14 chr15 chr16 chr17
chr18 chr20 chrY chr19 chr22 chr21 chrM" ," ")[[1]]
hg19Sort<-chromGenomeSort(sortedChroms)
bedData<-loadBedData("test.bed")
sortedBedData<-hg19Sort(bedData)
}

