% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifUtils.r
\name{compliment}
\alias{compliment}
\title{reverse compliment}
\usage{
compliment(string)
}
\arguments{
\item{string}{input base string, eg A[AG]CT}
}
\description{
Takes a string which can include base nucleaotides (AGCT) and
brakets ("[","]") and returns the reverse compliment including brakets
}
\examples{
compliment("[ACT]TA")
# [1] "TA[AGT]"
compliment(IUPACtoBase("STA"))
# [1] "TA[CG]"
consenusIUPAC(compliment(IUPACtoBase("STA")))
# [1] "TAS"
}

