% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{confint}
\alias{confint}
\title{Draw polygon confidence intervals}
\usage{
confint(x, ylow, yhigh, ...)
}
\description{
Draw polygon confidence intervals
}
\examples{

}
\keyword{function}
\keyword{helper}
