% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifUtils.r
\name{consenusIUPAC}
\alias{consenusIUPAC}
\title{Consensus IUPAC}
\usage{
consenusIUPAC(mstring)
}
\description{
The reverse of IUPACtoBase, takes base nucleotides and transforms
them them to their IUPAC equivelent
}
\examples{
> consenusIUPAC( IUPACtoBase("ARYS"))
[1] "ARYS"
}

