% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{contribution}
\alias{contribution}
\alias{contribution.default}
\title{Contribution Plot}
\usage{
contribution(prc, i, swapFun = function(string) string, swapColour = NULL,
  ...)

\method{contribution}{default}(prc, i, ...)
}
\arguments{
\item{prc}{The prc object to be ploted}

\item{i}{the PC to plot along}

\item{swapFun}{A function mapping the categoies of prc to another set of categories}

\item{swapColour}{A function that maps the results of swapFun to specifc colours}
}
\value{
A ggplot2 object
}
\description{
Generates the contribution histogram of the prc along a specified Principle Compoenent
}
\examples{

swapFun<-makeSwapfun("s1 sample1 s2 sample2 s3 sample3")
swapColour<-makeSwapfun("sample1 red sample2 blue sample3 orange")
contribution(CCCA_prc,swapFun,swapColour)
}

