% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifComparision.r
\name{distanceBetweenMotifs}
\alias{distanceBetweenMotifs}
\title{distanceBetweenMoitfs}
\usage{
distanceBetweenMotifs(sequence, motifs, locationsM, locationsC, m1, m2)
}
\description{
distanceBetweenMoitfs
}

