% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{E.dist.fit.f}
\alias{E.dist.fit.f}
\title{A generic distribution fitting function for the E variable. Used for random draws from E}
\usage{
E.dist.fit.f(E, densfun = densfun)
}
\arguments{
\item{E}{the time series of environment or ecosystem variable}

\item{densfun}{Either a character string or a function returning a density evaluated at its first argument.
Distributions "beta", "cauchy", "chi-squared", "exponential", "gamma", "geometric", "log-normal",
"lognormal", "logistic", "negative binomial", "normal", "Poisson", "t" and "weibull" are recognised,
case being ignored.}
}
\description{
A generic distribution fitting function for the E variable. Used for random draws from E
}
\examples{
E.dist.fit.f(rlnorm(1000,6,1.1),densfun="lognormal")
}
\keyword{climate}
\keyword{distribution,}
\keyword{environmental}
\keyword{normal}
\keyword{scenario}
\keyword{variable,}
