% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{Egamma.fit.f}
\alias{Egamma.fit.f}
\title{Fits a gamma distribution to the E variable. Used for random draws from E}
\usage{
Egamma.fit.f(E)
}
\arguments{
\item{E}{the time series of environment or ecosystem variable}
}
\description{
Fits a gamma distribution to the E variable. Used for random draws from E
}
\examples{
gamma.fit.f(rgamma(1000,rate=40,shape=200))
}
\keyword{climate}
\keyword{distribution,}
\keyword{environmental}
\keyword{gamma}
\keyword{scenario}
\keyword{variable,}
