% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.data.ccf.functions.r
\name{Egamma.plot.f}
\alias{Egamma.plot.f}
\title{Density of the gamma for different E accounting for mean and variance changes}
\usage{
Egamma.plot.f(Nrand, shape, rate, Emean.shift, E.variance)
}
\arguments{
\item{Nrand, }{the number of random values to make}
}
\description{
Density of the gamma for different E accounting for mean and variance changes
}
