% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{Eproj.list.f}
\alias{Eproj.list.f}
\title{creates a list of climate scenarios for developing CCF}
\usage{
Eproj.list.f(Emean.shifts, N, proj.years, Edist.a, Edist.b)
}
\arguments{
\item{N}{The number of different realisations of the future to create}

\item{proj.years}{The number of years to project into the future}

\item{mean.shift}{a vector of shifts in the mean of the gamma distribution of E}

\item{shape}{The gamma distribution shape parameter}

\item{rate}{The gamma distribution rate parameter}
}
\description{
creates a list of climate scenarios for developing CCF
}
\examples{
E.shifts= c(0,seq(-1,0.47,length=25)) #the first value must always be 0 (i.e. no mean shift)
Eprojection= Eproj.list.f(E.shifts, N=N, proj.years=proj.years, shape=gamma.shape, rate=gamma.rate)
}
\keyword{E}
\keyword{P/B,}
