% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{Eproj.np.f}
\alias{Eproj.np.f}
\title{Draws samples E variable for future climate scenarios - non-parametric option}
\usage{
Eproj.np.f(E, E.prediction.years, proj.years, N)
}
\arguments{
\item{E}{the time series of environment or ecosystem variable to resample (with replacement)}

\item{E.prediction.years}{the time series of environment or ecosystem variable}

\item{proj.years}{The number of years to project into the future}

\item{N}{The number of different realisations of the future to create}
}
\description{
Draws samples E variable for future climate scenarios - non-parametric option
}
\keyword{climate}
\keyword{environmental}
\keyword{non-parametric}
\keyword{projection,}
\keyword{scenario,}
\keyword{variable,}
