% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{Eprojgamma.f}
\alias{Eprojgamma.f}
\title{Draws samples from a gamma distribution fitted to E for future climate scenarios - parametric option}
\usage{
Eprojgamma.f(shape, rate, Emean.shift = 1, proj.years, N)
}
\arguments{
\item{shape}{The gamma distribution shape parameter}

\item{rate}{The gamma distribution rate parameter}

\item{Emean.shift}{The shift in gamma disribution to left or right to mimic mean shift in E variable (E units)}

\item{proj.years}{The number of years to project into the future}

\item{N}{The number of different realisations of the future to create}
}
\description{
Draws samples from a gamma distribution fitted to E for future climate scenarios - parametric option
}
\examples{
Eprojgamma.f(shape=262, rate=49.6, Emean.shift=1, proj.years=10, N=23)
}
\keyword{climate}
\keyword{environmental}
\keyword{non-parametric}
\keyword{projection,}
\keyword{scenario,}
\keyword{variable,}
