% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{Eprojnorm.f}
\alias{Eprojnorm.f}
\title{Draws samples from a gaussian distribution fitted to E for future climate scenarios}
\usage{
Eprojnorm.f(Edist.a, Edist.b, Emean.shift = 1, proj.years, N)
}
\arguments{
\item{Emean.shift}{The shift if the mean. This is done rather than changing the mean directly so it is generic for variety of distributions}

\item{proj.years}{The number of years to project into the future}

\item{N}{The number of different realisations of the future to create}

\item{E.dist.a}{The mean of the normal distribution}

\item{E.dist.b}{The standard deviation of the normal distribution}
}
\description{
Draws samples from a gaussian distribution fitted to E for future climate scenarios
}
\examples{
Eprojnorm.f(3, .05, Emean.shift=0, proj.years=10, N=23)
}
\keyword{climate}
\keyword{environmental}
\keyword{projection}
\keyword{scenario,}
\keyword{variable,}
