% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{EprojPB.f}
\alias{EprojPB.f}
\title{Sample the P/B values themselves rather than an environmental variable and then calculating P/B.}
\usage{
EprojPB.f(PBvec, PB.prediction.years, proj.years, N)
}
\arguments{
\item{PBvec}{the vector of PB values calculated from running the function PB.f}

\item{PB.prediction.years}{the years you want to sample from the PBvec (if 5th-10th year, 5:10)}

\item{proj.years}{The number of years to project into the future}

\item{N}{The number of different realisations of the future to create}
}
\description{
Sample the P/B values themselves rather than an environmental variable and then calculating P/B.
}
\keyword{climate}
\keyword{environmental}
\keyword{non-parametric}
\keyword{projection,}
\keyword{scenario,}
\keyword{variable,}
