% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.model.r
\name{F.strategy}
\alias{F.strategy}
\title{The fishing strategy as exploitation rate. F must be <=1}
\usage{
F.strategy(PB, years, moratorium = F)
}
\arguments{
\item{PB}{the data and model fit coming from applying the PB model (PB.f)}

\item{years}{the years whos mean will become the F for projections}

\item{moratorium}{if TRUE then there is no fishing going forward, otherwise calculated from relative F in specified years}
}
\description{
The fishing strategy as exploitation rate. F must be <=1
}
\keyword{Fishing}
\keyword{control}
\keyword{harvest}
\keyword{projections}
\keyword{rule,}
\keyword{strategy,}
