% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preferedDistance.r
\name{findMotifDist}
\alias{findMotifDist}
\title{find motif distance}
\usage{
findMotifDist(fasta, motifa, motifb, loc, ...)
}
\arguments{
\item{fasta}{The fasta sequences from Biostrings}

\item{motifa}{a string of IUPAC characters}

\item{motifb}{a string of IUPAC characters}

\item{loc}{a list of logicals equal in length to the fasta variable}

\item{...}{the distances at which peaks will be returned}
}
\value{
a list of indicies for peaks that have the motifs at prefered
distances
}
\description{
Identifies the indicies of the peaks that have two motifs at a prefered
distance from one another. Note that it only takes the smallest distance
into account when searching so this function is not suitible when the
expected distance between two motifs is great.
}

