% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{fishing.area3}
\alias{fishing.area3}
\title{Fishing area 3 - North Anticosti}
\format{A dataframe with four columns named. To be used by PBSmapping}
\usage{
data(fishing.area3)
}
\description{
A dataframe describing a polygon one of the three main turbot fishing areas in the Gulf - North Anticosti
}
\details{
\itemize{
  \item PID. The polygon id
  \item POS. The position
  \item X. Position of a longitude in decimal degrees. Western hemisphere is - degrees from the prime meridian
  \item Y. Position of a latitude in decimal degrees. Southern hemisphere is - degrees from the equator

}
}
\keyword{datasets}
