% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.model.r
\name{Fseq.f}
\alias{Fseq.f}
\title{The final B value at the specified time period at different Fs and climate scenario.}
\usage{
Fseq.f(PB, PBproj, Fseq, time.frame, N, K)
}
\arguments{
\item{PB}{the PB model output}

\item{Fseq}{a vector of F values to see how they perform}

\item{time.frame}{the time frame that you want to test make the test for. e.g. 5 years}

\item{N}{the number of monte carlo runs for projections}

\item{K}{the multiplier of maximum observed B (from Index.q) which will be the carrying capacity for the density dependent model}
}
\description{
The final B value at the specified time period at different Fs and climate scenario.
}
\examples{
Fout= Fseq.f(PB,Fseq=seq(0,.2,length=100),time.frame=time.frame, N=N, K=K)
}
\keyword{point,}
\keyword{probability}
\keyword{projection,}
\keyword{rank,}
\keyword{reference}
