% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rGREAT.r
\name{geneRanges}
\alias{geneRanges}
\title{Gene Ranges}
\usage{
geneRanges(chrom, tss, strand, proxUp, proxDown, distal, n = length(chrom))
}
\arguments{
\item{chrom}{A list of chromosomes, one for each gene}

\item{tss}{The list of gene TSS}

\item{strand}{The stream directions of each of the genes}

\item{proxUp}{Upstream basal extension (independent of other genes)}

\item{proxDown}{Downstream basal extension (independent of other genes)}

\item{distal}{The limit of gene extension (in either direction)}
}
\description{
Implements the two step Stanford GREAT genomic region generation.
First each gene is assigned a basal region defined as the TSS +/-
the proxUp and proxDown value (dependent on strand). Second each
Gene has its effective region expaned up to the lower of a distal
maximum and the basal limits of other genes.
The returned values are in the order of the input chrom tss strand
variables.
note: the indicies of the chrom tss and strand parameters should
all relate to the same gene, ie a chrom[100] tss[100] and strand[100]
should all be infromation from the same gene
}

