% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGREAT.r
\name{geneRangeToRegion}
\alias{geneRangeToRegion}
\title{Gene Range to Region}
\usage{
geneRangeToRegion(ranges, chrom, chroms = unique(chrom))
}
\arguments{
\item{ranges}{The genomic ranges for each gene}

\item{chrom}{A list of chromosomes, sequence should be shared with ranges}

\item{chroms}{a unique list of all chromomes in the crom list}
}
\value{
A list of conses. <chrom><start><end><gene 1>...<gene N>
}
\description{
Break the genome into regions based on the presence of genes. The
function relies on the output of geneRanges. It returnes mutualy
exclusive regions of the genome and the genes which relate to each
region.
}

