% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rGREAT.r
\name{genomicRegions}
\alias{genomicRegions}
\title{Genomic Regions}
\usage{
genomicRegions(chrom, tss, strand, proxUp, proxDown, distal,
  n = length(chrom), chroms = unique(chrom))
}
\arguments{
\item{chrom}{A list of chromosomes, one for each gene}

\item{tss}{The list of gene TSS}

\item{strand}{The stream directions of each of the genes}

\item{proxUp}{Upstream basal extension (independent of other genes)}

\item{proxDown}{Downstream basal extension (independent of other genes)}

\item{distal}{The limit of gene extension (in either direction)}
}
\description{
From a list of genes chromosomes tss and strandedness split the genome
into mutualy exclusive regions defined by the gene regions of influence
These regions can be used to relate ChIP binding data to genes. This
function combines the geneRagnges and geneRangeToRegion function into
a single form.
}
\examples{
heightFile<-"peaks/jan/combined_heights.bed"
geneFile<-"~/hg19.RefSeqGenes.csv"
geneList<-read.delim(geneFile)
chrom<-as.character(geneList$chrom)
tss<-as.numeric(geneList$txStart)
strand<-geneList$strand
# double check to make sure that levels(strand) > c("-","+")
levels(strand)<-c(-1,1)
strand<-as.numeric(strand)
statsAndData<-readSplitTable(heightFile)
regions<-genomicRegions(chrom,tss,strand,1000,5000,1000000)
}
\seealso{
geneRanges
}

