% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaUtils.r
\name{genSwapFun}
\alias{genSwapFun}
\title{genSwapFun}
\usage{
genSwapFun(input, output)
}
\arguments{
\item{input}{The input mapping}

\item{ouput}{The output mapping}
}
\value{
A function that maps each input to a single output
and returns a string
}
\description{
takes two equal length vectors and returns a function that maps
each input to its corresponding output. This can be used to alias
vectors if the mapping is known beforehand.
Note: the input vector must consist of strings.
}
\examples{
swapFun<-genSwapFun(c("a","b"),c(1,2))
swapFun("a")==1
swapFun("b")==2
swapFun("c")==NA
}

