% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{getBathy}
\alias{getBathy}
\title{Get bathymetric data and plot on the map}
\usage{
getBathy(bathyFname, isob = c(100, 200, 500, 800, 1200, 2400),
  minVerts = 3)
}
\description{
Get bathymetric data and plot on the map
}
\examples{

}
\keyword{function}
\keyword{helper}
