% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCCA.r
\name{getMotifInfo}
\alias{getMotifInfo}
\title{Get Motif Info}
\usage{
getMotifInfo(env, motifs)
}
\arguments{
\item{env}{}

\item{motifs}{The sequences being looked for}
}
\value{
env with the locations of motifs locations of peaks and
set of locations that are unique
}
\description{
Find the location of motifs for downstream analysis
}

