% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGREAT.r
\name{greatGeneAssoc}
\alias{greatGeneAssoc}
\title{Great Gene Assoc}
\usage{
greatGeneAssoc(bedData, genes, geneList)
}
\arguments{
\item{bedData}{The binding locations of interest}

\item{genes}{The genome split into gene regions}

\item{geneList}{A list with a 1-1 relation to the gene regions}
}
\value{
A filtered geneList
}
\description{
Compare peak locations to predetermined genomic  regions and return
a list of genes that corospond to the overlaped gene regions. This
function acts as a filter for a gene list whose indicies are shared
with the gene regions. Only the genes within regions that have peaks
are returned.
}
\examples{
bedFile<-"peaks/jan/combined.bed"
bed<-read.delim(bedFile)
geneFile<-"~/hg19.RefSeqGenes.csv"
geneList<-read.delim(geneFile)
chrom<-as.character(geneList$chrom)
tss<-as.numeric(geneList$txStart)
strand<-geneList$strand
# double check to make sure that levels(strand) > c("-","+")
levels(strand)<-c(-1,1)
strand<-as.numeric(strand)
statsAndData<-readSplitTable(heightFile)
regions<-genomicRegions(chrom,tss,strand,1000,5000,1000000)
genes<-greatGeneAssoc(bed,regions,geneList)
}

