% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/height.r
\name{hg19Sort}
\alias{hg19Sort}
\title{hg19Sort}
\source{
\url{https://github.com/alexjgriffith/mulcal}
}
\usage{
hg19Sort(data)
}
\arguments{
\item{data}{[chr,start,end]}
}
\value{
The same members of data but sorted on chr and start
}
\description{
Reorders the chro factor in data to that which is outputed by BWA.
This allows for the sorting step to be skipped, however if another chromasome is used then a new order must be defined.
}
\author{
Alexander Griffith griffita@gmail.com
}

