% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{intersectN}
\alias{intersectN}
\title{instersect N}
\usage{
intersectN(...)
}
\arguments{
\item{...}{a variable number of conses to be intersected}
}
\value{
a vector containing the intersecting values.
}
\description{
Recursivly intersects a list of vectors. Takes 0 to N vectors. 
The default behavior for a length of 0 is to return c(), i.e. NULL.
intersectN(a,b,c) ~ Reduce(intersect,list(a,b,c))
}
\examples{
intersectN()
# > NULL
intersectN(c(1,2,3))
# > 1 2 3 
intersectN(c(1,2,3),c(2,3,4))
# > 2 3
intersectN(c(1,2,3),c(2,3,4),c(3,4,5))
#> 3
}
\seealso{
\code{\link{unionN}\link{intersect}}
}
\keyword{intersect}
\keyword{recursive}

