% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{joinBathy}
\alias{joinBathy}
\title{Join bathymetric contours}
\usage{
joinBathy(CP, isobA = 100, isobB = 300, col = "grey")
}
\description{
Join bathymetric contours
}
\examples{

}
\keyword{function}
\keyword{helper}
