% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.data.ccf.functions.r
\name{Kobe.f}
\alias{Kobe.f}
\title{Make Kobe plots from simple data with overlay of another time series}
\usage{
Kobe.f(PB, E, ...)
}
\arguments{
\item{PB}{the data and model fit coming from applying the PB model (PB.f)}

\item{E}{the time series to overlay and colour code Kobe years}

\item{...}{par options for plot}
}
\description{
Make Kobe plots from simple data with overlay of another time series
}
\examples{
Kobe.f(PB=PB,E=PB$E)
}
\keyword{Bmsy,}
\keyword{Kobe,}
\keyword{environment}
\keyword{point,}
\keyword{reference}
