% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{lmP}
\alias{lmP}
\title{functions to extract P fit diagnostics from an lm object}
\usage{
lmP(lm.fit.object)
}
\description{
functions to extract P fit diagnostics from an lm object
}
\examples{

}
\keyword{function}
\keyword{helper}
