% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{lnorm.fit.f}
\alias{lnorm.fit.f}
\title{Fits a lognormal distribution to the E variable. Used for random draws from E}
\usage{
lnorm.fit.f(E)
}
\arguments{
\item{E}{the time series of environment or ecosystem variable}
}
\description{
Fits a lognormal distribution to the E variable. Used for random draws from E
}
\examples{
lnorm.fit.f(rlnorm(1000,2.56,0.055))
}
\keyword{climate}
\keyword{distribution,}
\keyword{environmental}
\keyword{lognormal}
\keyword{scenario}
\keyword{variable,}
