% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadPWM.r
\name{loadPWM}
\alias{loadPWM}
\title{Load PWM}
\usage{
loadPWM(fileLocation, version = "homer")
}
\arguments{
\item{fileLocation}{The location of the motif file}

\item{version}{the name of the program that outputed the file. Currently the only versions are homer and jaspar.}
}
\value{
a unified pwm
}
\description{
This utiltiy is designed to load several PWM formats and create a uniform layout within R for analysis
}

