% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/motifComparision.r
\name{locHist2}
\alias{locHist2}
\title{locHist}
\usage{
locHist2(t1, x, xlab = "Histogram", limits = c(-32, 32))
}
\arguments{
\item{t1}{}

\item{x}{}

\item{xlab}{}

\item{limits}{}
}
\description{
locHist
}

