% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAFS.r
\name{makeAFS}
\alias{makeAFS}
\title{Make AFS}
\usage{
makeAFS(peakList, categories, format = "xls", pvalue = NULL, width = 700)
}
\arguments{
\item{peakList}{A list of strings refering to file locations containing
Peak data}

\item{categories}{The monikers these peak files will be refered to ass}

\item{format}{The format to be imported, either bed or xls}

\item{pvalue}{The pvalue cut off to be applied, only valid with xls format}

\item{width}{The min width summits must be from one another in order to
cluster}
}
\value{
The AFS generated from the input peak lists <chr><summit><catagory>
}
\description{
Take a list of peaks and create a unified set. The unified set is
created by grouping summits that fall within \code{width} of one
another. The mean of the grouped summits is used as a shared summit
between the grouped peaks. The start and end of the new peak is defined
as the new sumit +/ 1/2 \code{width}.
}
\examples{
peakList<-sapply(c("sample1.xls","sample2.xls","sample3.xls"),
  function(file){
    system.file("extdata", file, package = "CCCA")
  })
categories<-c("s1","s2","s3")
sampleafs<-makeAFS(peakList,categories,pvalue=20)
}
