% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccca.r
\name{makeCCCA}
\alias{makeCCCA}
\title{makeCCCA}
\usage{
makeCCCA(dataSets, peakLists, categories, pvalue)
}
\arguments{
\item{dataSets}{list of files containing the raw reads in bed format}

\item{peakLists}{list of files containing the peaks reads in MACS xls format}

\item{categories}{the moniker to be attached to each of the data set files}

\item{pvalue}{the pvalue that the AFS peaks will be filtered by}
}
\value{
a list of class ccca containing afs,udm, and prc
}
\description{
Generate a ChIP Component object from peaks, reads, and a list of categories
}
\examples{
dataSets<-sapply(c("raw_sample1.bed","raw_sample2.bed","raw_sample3.bed"),
  function(file){
    system.file("extdata", file, package = "CCCA")
  })
peakLists<-sapply(c("sample1.xls","sample2.xls","sample3.xls"),
  function(file){
    system.file("extdata", file, package = "CCCA")
  })
categories<-c("s1","s2","s3")
ccca<-makeCCCA(dataSets,peakLists,categories,20)
## Find dimensions that seperate data sets of interest
plot(ccca,1,2)
plot(ccca,1,3)
ccca<-local({
  nm<-CCCA::normalize(ccca$prc)
  ccca<-addReg(ccca,"s1",nm[,1]<(mean(nm[,1])-3*sd(nm[,1])))
  ccca<-addReg(ccca,"s2",nm[,1]>(mean(nm[,1])+3*sd(nm[,1])))
  ccca<-addReg(ccca,"s3",nm[,2]>(mean(nm[,2])+3*sd(nm[,2])))
  ccca<-addReg(ccca,"s1.me",ccca$reg[,"s1"] & !(ccca$reg[,"s2"] | ccca$reg[,"s3"]))
  ccca<-addReg(ccca,"s2.me",ccca$reg[,"s2"] & !(ccca$reg[,"s1"] | ccca$reg[,"s3"]))
  ccca<-addReg(ccca,"s3.me",ccca$reg[,"s3"] & !(ccca$reg[,"s2"] | ccca$reg[,"s1"]))
  ccca
})
\dontrun{
library(BSgenome.Hsapiens.UCSC.hg19)
genome=BSgenome.Hsapiens.UCSC.hg19
addFasta(ccca,genome)
}
}

