% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{makeLogic}
\alias{makeLogic}
\title{Make Logical}
\usage{
makeLogic(loc, size)
}
\arguments{
\item{loc}{a list of indicies}

\item{size}{the length of the output vector}
}
\value{
a vector of logicals
}
\description{
transforms a vector of indicies into logical values.
Requires the full size of the output list
}
\examples{
i<-c(1,2,3)
all(makeLogic(i,4) == c(TRUE,TRUE,TRUE,FALSE))
}

