% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSwapFun.r
\name{makeSwapFun}
\alias{makeSwapFun}
\title{string to swap fun}
\usage{
makeSwapFun(x)
}
\arguments{
\item{x}{input string}
}
\value{
a function mapping the even values of x to the odd
}
\description{
takes a string and produces a swap fun from it. The string is
broken allong spaces.
}
\examples{
makeSwapFun("a b c d")("b")=="a"
}
