% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeLatexTable.R
\name{makeTable}
\alias{makeTable}
\title{make table}
\usage{
makeTable(fun = defaultFun)
}
\description{
makes a latex table from a r matrix
}
\examples{
a<-matrix(c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18),
          ncol=9,dimnames=list(list("alex",2),
          list("a","b","c","d","e","f","g","h","i") ))
# > a
#      a b c d  e  f  g  h  i
# alex 1 3 5 7  9 11 13 15 17
# 2    2 4 6 8 10 12 14 16 18
cat(makeTable()(a,"llllllll",label="tab:centers",caption="data table"))

defaultFun<-function(data,layout,cn=" ",caption=NULL,label=NULL){
  rownames<-rownames(data)
  colnames<-colnames(data)
  ins<-Map(function(...) unlist(list(...)),rownames,apply(data,1,list))
  wf<-.p(tags,"table",opts="h")
  if(! is.null(caption))
    wf<-.c(wf,.p(singleTag,"caption",caption))
  if(! is.null(label))
    wf<-.c(wf,.p(singleTag,"label",label))
  wf<-.c(wf,
  .p(tags,'center'),
  .p(tags,"tabular",opts=layout,n=c("{","}")))
  wf(paste(makeLines(ins,paste(vb(),
                     vb(makeLine(unlist(list(cn,colnames)))))),vb()))
}
makeT<-makeTable(defaultFun)
cat(makeT(a,"llllllll",label="tab:centers",caption="data table"))
}

