% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{map.f}
\alias{map.f}
\title{Wrapper for the PBSmapping plotMap function}
\usage{
map.f(longs = c(-74, -50), lats = c(43, 52.25),
  land.colour = "sienna3", sea.colour = "lightblue")
}
\description{
Wrapper for the PBSmapping plotMap function
}
\examples{

}
\keyword{function}
\keyword{helper}
