% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{map.inset.f}
\alias{map.inset.f}
\title{An inset map for the main map function}
\usage{
map.inset.f(
  longs = c(-71, -50),
  lats = c(43, 52.25),
  land.colour = "sienna3",
  sea.colour = "lightblue"
)
}
\description{
An inset map for the main map function
}
\examples{

}
\keyword{function}
\keyword{helper}
