% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{mapziplist}
\alias{mapziplist}
\title{map zip list}
\usage{
mapziplist(...)
}
\arguments{
\item{...}{an arbitraraly long set of equal length lists}
}
\description{
Takes a list of lists and groups the nth member of each list
into a new one.
}
\examples{
mapziplist(list(1,2,3),list("A","B","C"))
# [[1]]
# [1] "1" "A"
# 
# [[2]]
# [1] "2" "B"
# 
# [[3]]
# [1] "3" "C"
}

