% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{mLength}
\alias{mLength}
\title{m Length}
\usage{
mLength(mat)
}
\arguments{
\item{mat}{input matrix}
}
\value{
the width of the matrix
}
\description{
Determines the length of a matrix or vector.
}
\examples{
mLength(c(1,2,3)) # > 3
mLength(cbind(c(1,2,3),c(4,5,6))) # > 3
}
\seealso{
\code{\link{mLength}}
}

