% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{modulous}
\alias{modulous}
\title{modulous}
\usage{
modulous(x, m)
}
\arguments{
\item{x}{numerator}

\item{m}{denominator}
}
\value{
an numeric value representing the modlulous
}
\description{
Turns the %% infix operater into a function
}

