% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{mor}
\alias{mor}
\title{method or}
\usage{
mor(...)
}
\arguments{
\item{...}{variable length list of inputs}
}
\value{
the first non NULL value
}
\description{
iterates through the members of the input vector and returns the first
variable that has a non NULL value
}
\examples{
mor() # > NULL
mor(1,2,3) # > 1
mor(NULL,1) # > 1
# useful for generalizing dim(x)[2]
mor(dim(c(1,2,3))[2],1) #> 1
}

