% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{multiPlot}
\alias{multiPlot}
\title{Multi Plot}
\usage{
multiPlot(plots, rows = 1, cols = 1)
}
\arguments{
\item{plots}{A list of plots to be ploted}

\item{rows}{The number of rows to plot}

\item{cols}{The number of columns to plot}
}
\description{
Takes mutliple ggplots and aranges them using grid.layout. The
result can be saved using standard png and pdf commands.
}
\examples{
a<-data.frame(a=runif(100),b=runif(100),x=seq(100))
p<-c(ggplot(a,aes(x=x,y=a))+geom_points(),
     ggplot(a,aes(x=x,y=b))+geom_points())
multiPlot(p,2)
}

