% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{mWidth}
\alias{mWidth}
\title{m Width}
\usage{
mWidth(mat)
}
\arguments{
\item{mat}{input matrix}
}
\value{
the width of the matrix
}
\description{
Returns the width of a vector, if it is a non matrix
it returns 1 rather than NULL.
}
\examples{
mWidth(c(1,2,3)) # > 1
mWidth(cbind(c(1,2,3),c(4,5,6))) # > 2
}
\seealso{
\code{\link{mLength}}
}

