% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{ncod}
\alias{ncod}
\title{Northern cod data}
\format{
A dataframe with four essential columns named (Year, Index, Catch, E) and many other E variable columns.
}
\source{
Regular, P. 2019. Northern cod assessment res doc and dash board. CSAS in press

       NOAA Weather Service. 2019. https://www.cpc.ncep.noaa.gov/products/precip/CWlink/pna/nao.shtml
}
\usage{
data(ncod)
}
\description{
Northern cod (2J3KL) data. The survey index is not used here but the output of Regular's extended northern cod model. Therefore
there is no need for a q different than 1. The environmental indices are all from NOAA (https://www.cpc.ncep.noaa.gov/products/precip/CWlink/pna/nao.shtml)
}
\details{
\itemize{
  \item Year. The year
  \item Index. The survey swept area biomass of shrimp (kt). Also male and female spit bionmass.
  \item Catch. The reported catch of the stock (kt)
  \item E. The E variable you choose to be your climate variable (probably one of the variables below)
  \item X. Various oscillation indices. Annual averages of daily values.

}
}
\keyword{datasets}
