% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{newpar.f}
\alias{newpar.f}
\title{Relist parameter objects from a particular projection}
\usage{
newpar.f(skeleton)
}
\arguments{
\item{skeleton}{the name attribute from the params list}
}
\description{
Relist parameter objects from a particular projection
}
\examples{
skel= attr(params,"names")
my.projection.params= newpar.f(skeleton=skel)
save(my.projection.params, file="my.projection.params.rda")
}
\keyword{list}
\keyword{parameters,}
\keyword{projection,}
