% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{norm.fit.f}
\alias{norm.fit.f}
\title{Fits a normal distribution to the E variable. Used for random draws from E}
\usage{
norm.fit.f(E)
}
\arguments{
\item{E}{the time series of environment or ecosystem variable}
}
\description{
Fits a normal distribution to the E variable. Used for random draws from E
}
\examples{
lnorm.fit.f(rnorm(1000,2.65,0.703))
}
\keyword{climate}
\keyword{distribution,}
\keyword{environmental}
\keyword{normal}
\keyword{scenario}
\keyword{variable,}
