% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.data.ccf.functions.r
\name{norm.plot.f}
\alias{norm.plot.f}
\title{Density of the gamma for different E accounting for mean and variance changes}
\usage{
norm.plot.f(Nrand, Edist.a, Edist.b, Emean.shift, E.var.inc)
}
\arguments{
\item{Nrand, }{the number of random values to make}
}
\description{
Density of the gamma for different E accounting for mean and variance changes
}
