% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{normalize}
\alias{normalize}
\alias{normalize.default}
\title{Normalize Plot}
\usage{
normalize(x, ...)

\method{normalize}{default}(x, ...)
}
\arguments{
\item{x}{Vector or matrix to be normalized}
}
\description{
Normalize Plot
}
\examples{
r<-runif(100,0,100)
normalize(x)
b=NULL
b$eivenVector<-matrix(r,10)
class(b)<-"prc"
normalize(b)$eigenVector
}

