% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{ocBathy}
\alias{ocBathy}
\title{Eastern Canada Bathymetry}
\format{
A list
}
\usage{
data(ecan.bathy)
}
\description{
A large list with x, y, z elements for plotting isobaths on the map
}
\details{
\itemize{
  \item X. Position of a longitude in decimal degrees. Western hemisphere is - degrees from the prime meridian
  \item Y. Position of a latitude in decimal degrees. Southern hemisphere is - degrees from the equator
  \item Z. The depth in m at the x, y point

}
}
\keyword{datasets}
