% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcaUtils.r
\name{orderBed}
\alias{orderBed}
\title{Order Bed}
\usage{
orderBed(ret)
}
\arguments{
\item{ret}{data frame}
}
\value{
sorted data frame
}
\description{
sorts a data.frame based on the first and third columns
}
\examples{
orderBed(data.frame(c("a","a","c"),c(1,2,3),c(5,4,6)))
}

