% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{overlap}
\alias{overlap}
\title{Overlap}
\usage{
overlap(f)
}
\arguments{
\item{f}{n by m logical matrix}
}
\value{
A m by m matrix of column overlaps
}
\description{
Takes a matrix of logical inputs, each column representing a data set each
row representing a location on the genome,
}
\examples{
lmatrix<-matrix(as.logical(c(1,1,1,1,1,1,0,0,1,0,1,0)),ncol=3)
overlap(lmatrix)
}

