% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.model.r
\name{P.R.for.EF.f}
\alias{P.R.for.EF.f}
\title{For multiple E and F scenario, the probability of achieving the refernece point is determined}
\usage{
P.R.for.EF.f(E.CCF, PB.CCF, Fs, PB, ref.pt, Bstart.mult, K, theta = 1)
}
\arguments{
\item{E.CCF}{The list of E scenarios}

\item{PB.CCF}{the list of PB scenarios from E.CCF and PvsE relationship}

\item{Fs}{The vector of Fs to test}

\item{PB}{the data and model fit coming from applying the PB model (PB.f)}

\item{ref.pt}{there reference point multiplier as a proportion of the reference period}

\item{Bstart.mult}{the starting biomass as a proportion of the last data year biomass (Index.q)}

\item{K}{The multiplier of maximum observed biomass to be carrying capacity}

\item{theta}{the skewness of the density dependence factor (1=Schaeffer)}
}
\description{
For multiple E and F scenario, the probability of achieving the refernece point is determined
}
\examples{
fs= c(seq(0,.1,length=10),seq(0.11,0.5,length=10))
P.R.for.EF.f(E.CCF=ECCF,PB.CCF=PBCCF,Fs=fs, PB=PB, ref.pt=ref.pt, Bstart.mult=Bstart.mult, K=K, theta=theta)
}
\keyword{point,}
\keyword{probability}
\keyword{projection,}
\keyword{rank,}
\keyword{reference}
