% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifUtils.r
\name{pairwiseCombs}
\alias{pairwiseCombs}
\title{pariwise Combinations}
\usage{
pairwiseCombs(options)
}
\arguments{
\item{options}{list or cons}
}
\value{
cbind(lista,listb)
}
\description{
Takes a list and returns the pairwise combinations of all members
}
\examples{
characters<-c("A","C","G","T")
pairwiseCombs(characters)
}

