% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.documentation.r
\docType{data}
\name{params}
\alias{params}
\title{Projection parameters}
\format{A list with 19 elements}
\usage{
data(params)
}
\description{
A list with all the default projection parameters needed to run a projection. You need to unlist this
so that each named element (level 1) in the list is an object in the global environment. You can change
these unlisted objects individually. It can be a good idea to unlist the basic parameters file before
each projection and then changing only the things you want to make sure you are always starting from
the same point each time.
}
\details{
\itemize{
  \item q. Survey catchability. Probably 1 or less (scalar)
  \item ref.pt. A multiplier of the biomass in the reference period to estimate the reference point (scalar)
  \item risk. The acceptable risk of not achieving the reference point in the specified time.frame (scalar 0--1).
  \item time.frame. The acceptable time period for achieving the reference point objective (scalar)
  \item Bstart.mult. The multiplier of the last data year's biomass which will start the projection (scalar)
  \item K. Multiplier of the maximum observed biomass that will be considered the carrying capacity for the density dependent model (scalar)
  \item theta. The degree of asymmetry in the density dependent production curve (scalar, 1=symmetric).
  \item model.type. This is the kind of model you want to use to fit the PB vs E relationship: gam.adaptive" "poly", "gam.adaptive"
  \item poly.degree. The degree of the polynomial to fit for P/B vs E if you chose "poly" for model type
  \item knots. The number of knots for a gam.adaptive fit. A good value is n/2.5
  \item add.resids. If 1 then multiplicative residuals are sampled from P/B vs E curve. 0 if no residuals (scalar values other than 0 and 1 should NEVER be used)
  \item Emean.shift. The change in the mean of the E distribution for a future climate scenario (scalar, + or -)
  \item E.var.inc. The change in the variance of the E distribution for a future climate scenario (scalar, 1 for as is, >1 decreases variance)
  \item N. The number of Monte Carlo samples for the projection (scalar)
  \item proj.years. The number of years to project into the future. Must be >= time.frame (scalar)
  \item N.CCF. The number of Monte Carlo samples for each projection for very unique commbination of Emean.shift and F into the future (scalar)
  \item Emean.shifts. The Emean.shifts to try for determining CCFs (vector)
  \item fs. The exploitation rates to try for determining CCFs (vector)

}
}
\keyword{datasets}
