% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{pass}
\alias{pass}
\title{Pass}
\usage{
pass(x)
}
\arguments{
\item{x}{variable}
}
\value{
x unchanged
}
\description{
This function is used as a placeholder when a filter is required.
}
\examples{
a<-1
pass(a) #> 1
}

