% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.model.r
\name{PB.f}
\alias{PB.f}
\title{Determine the annual P/B of the population. Write input file with P/B column}
\usage{
PB.f(dataset, ref.years, q)
}
\arguments{
\item{dataset}{the dataframe with at least the columns named "Year", "Catch", "Index", "E".
There cannot be missing years but the code does not check for that - beware!}

\item{ref.years}{the years which are considered the reference point year (e.g. for B, E, catch). It is up to you
think it is representative of Blim, Bmsy, Busr or whatever. Just beware that reference point multipliers will be
based on it in other areas of the simulation. So if it is a Bmsy proxy then the Blim ref.pt input logically might be about 0.4.}

\item{q}{the survey catchability}
}
\description{
Determine the annual P/B of the population. Write input file with P/B column
}
\examples{
PB= PB.f(dataset,1)
}
\keyword{P/B,}
\keyword{capacity,}
\keyword{carrying}
\keyword{catcahbility,}
\keyword{catch}
\keyword{commercial}
\keyword{environmental}
\keyword{survey}
\keyword{variable,}
