% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{PB.for.projection.f}
\alias{PB.for.projection.f}
\title{A function calculated P/B for projection based on the climate scenario. It samples the residuals of the fitted relationship and adds one to each draw}
\usage{
PB.for.projection.f(PvsE, Eproj, add.residuals = add.resids)
}
\arguments{
\item{PvsE}{The fitted PB vs E model object}

\item{Eproj}{The matrix of E values for the projection. From running Eprojgamma.f}

\item{add.residuals}{if yes=1, no=0. DO NOT USE ANY OTHER VALUE!! This formulation is faster because it avoids an if statement}
}
\description{
A function calculated P/B for projection based on the climate scenario. It samples the residuals of the fitted relationship and adds one to each draw
}
\keyword{E}
\keyword{P/B,}
