% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.model.r
\name{PBE.fit.f}
\alias{PBE.fit.f}
\title{Fit a PB vs E relationship}
\usage{
PBE.fit.f(PB, model.type, knots, poly.degree)
}
\arguments{
\item{PB}{the data and model fit coming from applying the PB model (PB.f)}

\item{model.type}{the kind of model to fit ("poly", "gam", "gam.adaptive","avg")}

\item{knots}{the number of knots for adaptive GAM}

\item{poly.degree}{the degree of the polynomial to fit}
}
\description{
The various model fits: polynomial "poly", GAM "gam", adaptive GAM "gam.adaptive"
    and resamples from the PB values "avg" can be chosen. avg just fits a linear model with slope = 0
    and then resamples the residuals which is effectively the same as just sampling the P/B values directly,
    i.e. it does not force a relationship between P/B and E and therefore the future is just a resampling
    of the past.
}
\keyword{E}
\keyword{P/B,}
\keyword{line,}
\keyword{trend}
