% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{PBproj.list.f}
\alias{PBproj.list.f}
\title{Creates the PB scenario based on the climate (E) and the P/B vs E relationship}
\usage{
PBproj.list.f(PvsE, Eprojection)
}
\arguments{
\item{PvsE}{the fitted model object of P/B vs E}

\item{Eprojection}{the list of climate scenarios created by running Eproj.list.f}
}
\description{
Creates the PB scenario based on the climate (E) and the P/B vs E relationship
}
\examples{
PBprojection= PBproj.list(PvsE=PvsE, Eprojection=Eprojection)
}
\keyword{E}
\keyword{P/B,}
