% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.r
\name{pca}
\alias{pca}
\title{PCA}
\usage{
pca(data, norm = "qn")
}
\arguments{
\item{data}{The matrix to be analyzed}

\item{norm}{The normalization method
norm may either be a user provided function or one of the
following
\describe{
\item{rowSumOne}{ all rows sum to 1}
\item{colSumOne}{ all columns sum to 1}
\item{normRow}{all rows have mean 0 and sd 1}
\item{normCol}{all cols have mean 0 and sd 1}
\item{rowsVarOne}{ all rows have sd 1}
\item{colsVarOne}{ all cols have sd 1}
\item{qn}{ quantile normalization}
\item{none}{ no normalization}
}}
}
\value{
list($normData,$eigenVectors)
}
\description{
Normalizes a data matrix and then preformes quantile normalization
using \code{prcomp} on the transform of the normalized data. The
eigenvectors are returned in a list with the normalized data
}

