% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pca.r
\name{peakPartitions}
\alias{peakPartitions}
\title{peak partitions}
\usage{
peakPartitions(pcs, l = mWidth(pcs), tests = rep("gt", l), ns = rep(1, l))
}
\arguments{
\item{pcs}{the vectors to be assesed}

\item{l}{the number of vectors being assesed}

\item{tests}{the tests being applied to each vector}

\item{ns}{the number of stadard divations from the mean is considered
significant}
}
\description{
Partitions a vector of weights based on its mean and standard deviation.
}

