% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{plotPCs}
\alias{plotPCs}
\title{plot PCs}
\usage{
plotPCs(pcs, pos, data, cats, lab = c("xlabel", "ylable", "Title"))
}
\arguments{
\item{pcs}{the eiginvectors}

\item{pos}{a list indicating which two dimensions should be ploted}

\item{data}{the normalized data}

\item{cats}{the catagories, used to label the plot}

\item{lab}{the labeling information}
}
\description{
A tool used to plot the cross product of the normalized data
and the eigenvectors, it relies only on the standard library
ploting function
}
\examples{
heightFile<-"~/Dropbox/UTX-Alex/jan/combined_heights.bed"
catFile<-"~/Dropbox/UTX-Alex/jan/catagories"
cats<-as.character(unlist(read.table(catFile)))
score<-loadHeightFile(heightFile)$data
normData_eigen<-pca(score)
plotPCs(normData_eigen[[2]],c(1,3),normData_eigen[[1]],cats)
}

