% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.model.r
\name{PofF.f}
\alias{PofF.f}
\title{Calculates the probability of being at or above the reference B value at the end of the time frame specified.}
\usage{
PofF.f(PB, Fprob, ref.pt)
}
\arguments{
\item{PB}{the PB model output}

\item{Fprob}{the output of Fseq.f, i.e. the final B value at the end of a specified period}

\item{ref.pt}{the multiplier of the reference period giving the reference point}
}
\description{
Calculates the probability of being at or above the reference B value at the end of the time frame specified.
}
\examples{
PofF.f(PB,Fout)
}
\keyword{point,}
\keyword{probability}
\keyword{projection,}
\keyword{rank,}
\keyword{reference}
