% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.model.r
\name{projection.f}
\alias{projection.f}
\title{Performs the projection with climate scenario and fishing strategy}
\usage{
projection.f(PB, Bstart.mult, PBproj, Fstrat, K, theta = 1)
}
\arguments{
\item{PB}{the data and model fit coming from applying the PB model (PB.f)}

\item{Bstart.mult}{the proporption of the last data year's biomass used to start the projection}

\item{PBproj}{The matrix of projected PB values based on climate scenario}

\item{Fstrat}{The fishing strategy}

\item{K}{The multiplier of maximum observed biomass to be carrying capacity}

\item{theta}{the skewness of the density dependence factor (1=Schaeffer)}
}
\description{
Performs the projection with climate scenario and fishing strategy
}
\examples{
Bproj= projection.f(PB=PB, Bstart.mult=Bstart.mult, PBproj=PBproj, Fstrat, K=K, theta=1)
}
\keyword{change,}
\keyword{climate}
\keyword{density}
\keyword{dependence}
\keyword{fishing,}
\keyword{production,}
\keyword{projection,}
