% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate.scenario.functions.r
\name{PvsE.resids.f}
\alias{PvsE.resids.f}
\title{A function random sample of residuals for PB vs E relationship, creates additive terms for the P/B samples}
\usage{
PvsE.resids.f(PvsE, proj.years, N)
}
\arguments{
\item{PvsE}{The fitted PB vs E model object}

\item{proj.years}{The number of years to project into the future}

\item{N}{The number of different realisations of the future to create}
}
\description{
A function random sample of residuals for PB vs E relationship, creates additive terms for the P/B samples
}
\keyword{E}
\keyword{P/B,}
