% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{qn}
\alias{qn}
\title{Quantile Normalization}
\usage{
qn(data)
}
\arguments{
\item{data}{matrix of data to be normalized}
}
\value{
matrix of the same dimensions as data
}
\description{
Applies quantile normalization to data in matrix form.
}
\details{
\itemize{
\item{1. Order each of the columns}
\item{2. Sum each orderd row and normalize by the width}
\item{3. Put the data back in order}
}
}
\examples{
matr<-cbind(rbind(5,2,3,4),rbind(4,1,4,2),rbind(3,4,6,8))
qn(matr)
}

