% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCF.model.r
\name{rankprob.f}
\alias{rankprob.f}
\title{Determines the probability of achieving objective (reference point) each year for each monte carlo run}
\usage{
rankprob.f(proj.out, PB, ref.pt)
}
\arguments{
\item{proj.out}{the output of the monte carlo projections (generated by running projection.f)}

\item{PB}{the PB model output}

\item{ref.pt}{the multiplier of the reference period biomass to calculate reference point of interest}
}
\description{
Determines the probability of achieving objective (reference point) each year for each monte carlo run
}
\keyword{point,}
\keyword{probability}
\keyword{projection,}
\keyword{rank,}
\keyword{reference}
