% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAFS.r
\name{readAFS}
\alias{readAFS}
\title{Read AFS
Read an AFS file}
\usage{
readAFS(fname)
}
\arguments{
\item{fname}{Save filename}
}
\value{
An AFS object
}
\description{
Read AFS
Read an AFS file
}
\examples{
filename<-system.file("extdata","sample.afs", package = "CCCA")
sampleAFS<-readAFS(filename)
}
