% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcaUtils.r
\name{readCategories}
\alias{readCategories}
\title{Read Categories}
\usage{
readCategories(...)
}
\arguments{
\item{filename}{the filename}
}
\value{
a vector of strings
}
\description{
read a single column file and return a vector of strings
}

