% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPeaksbed.r
\name{readPeaksBed}
\alias{readPeaksBed}
\title{Read Peaks BED format}
\usage{
readPeaksBed(file, name = file)
}
\arguments{
\item{file}{The MACS xls output to be imported}

\item{name}{A string as a moniker for this data set}
}
\value{
a data.frame <chr><summit><name> | NULL if file has no entries
}
\description{
Read in peaks that MACs has output in its bed format. Works with MACS 2.1 and 1.3.1.
Cannot filter peaks based on macs score.
}
\examples{
# need to create sample.xls
S1_20<-readPeaksBed("sample.bed","s1",20)
}

