% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPeaksXLS.r
\name{readPeaksXLS}
\alias{readPeaksXLS}
\title{Read Peaks XLS format}
\usage{
readPeaksXLS(file, name = file, pvalue = 0)
}
\arguments{
\item{file}{The MACS xls output to be imported}

\item{name}{A string as a moniker for this data set}

\item{pvalue}{log10 pvalue used as a cut off, default 0}
}
\value{
a data.frame <chr><summit><name> | NULL if file has no entries
}
\description{
Read in peaks that MACs has output in its XLS format. Works with MACS 2.1 and 1.3.1
}
\examples{
# need to create sample.xls
filename<-system.file("extdata","sample1.xls", package = "CCCA")
readPeaksXLS(filename)
}

