% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUDM.r
\name{readUDM}
\alias{readUDM}
\title{Read UDM}
\usage{
readUDM(fname)
}
\arguments{
\item{fname}{Save filename}
}
\value{
A udm object
}
\description{
Read the UDM from its tab format
}
\examples{
filename<-system.file("extdata","sample.udm", package = "CCCA")
sampleUDM<-readUDM(filename)
}

