% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CCCA.r
\name{sepAxis}
\alias{sepAxis}
\title{seperate axis}
\usage{
sepAxis(..., inclusive = TRUE)
}
\arguments{
\item{...}{list of options list(name,fun,sd,pc)}

\item{inclusive}{if false the subsets are mutualy exclusive}
}
\description{
Generates a function that can be used to isolate subsets of peaks using
the eigenVectors of the normalized data.
}

