% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaUtils.r
\name{simpleSwapFun}
\alias{simpleSwapFun}
\title{simple swap fun}
\usage{
simpleSwapFun(char)
}
\arguments{
\item{char}{input string}
}
\value{
a function mapping the odd values of x to the even
}
\description{
this function is proforms the same role of stringToSwap, but the even and
odd strings are reversed.
}
\examples{
simpleSwapFun("a b c d")("a")=="b"
}

