% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.r
\name{sortDataFrame}
\alias{sortDataFrame}
\title{sort data frame}
\usage{
sortDataFrame(dd, ...)
}
\description{
A small utility to functionalize the sorting of data frames
}
\examples{
dd<-data.frame(initial1=LETTERS[runif(100,1,24)],initial2=LETTERS[runif(100,1,24)],age=floor(runif(100,21,35)))
sortDataFrame(dd,"initial2","age")
}

