% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{splitZip}
\alias{splitZip}
\title{Split Zip}
\usage{
splitZip(inList)
}
\description{
Extracts the first and second values from a list of pairs.
}
\examples{
splitZip(createZip(c(1,2,3,4,5,6))) == list(c(1,3,5),c(2,4,6))
}

