% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{stem}
\alias{stem}
\title{stem}
\usage{
stem(x, y, pch = 16, linecol = 1, clinecol = 1, ...)
}
\description{
Produces a stem plot similar to matlab using the plotting utility.
}
\examples{
stem(1:10,runif(10))
}
\author{
Matti Pastell
}
\references{
\url{http://www.r-bloggers.com/matlab-style-stem-plot-with-r/}
}
\seealso{
\code{\code{plot}}
}

