% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{unionN}
\alias{unionN}
\title{union N}
\usage{
unionN(...)
}
\arguments{
\item{...}{a variable number of vectors to have their union taken}
}
\value{
a vector containing the union of the values.
}
\description{
Recursivly tales the union of a list of vectors. Takes 0 to N vectors. 
The default behavior for a length of 0 is to return c(), i.e. NULL.
unionN(a,b,c) ~ Reduce(union,list(a,b,c))
}
\examples{
intersectN()
# > NULL
intersectN(c(1,2,3))
# > 1 2 3 
intersectN(c(1,2,3),c(2,3,4))
# > 2 3
intersectN(c(1,2,3),c(2,3,4),c(3,4,5))
#> 3
}
\seealso{
\code{\link{intersectN} \link{union}}
}
\keyword{recursive}
\keyword{union}

