% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/afs.r
\name{unityOutput}
\alias{unityOutput}
\title{unity output}
\usage{
unityOutput(peaks, intChr, intSummit, intname)
}
\arguments{
\item{peaks}{A 2xN matrix representing overlaping peak regions}

\item{intChr}{N length list of chomosomes represented as integers}

\item{intSummit}{The summit location}

\item{intname}{The data set or name that the peak come froms in int form}
}
\description{
Wrapper for the unityOutput C function.
}

