% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAFS.r
\name{writeAFS}
\alias{writeAFS}
\title{Write AFS
Write the AFS to a table}
\usage{
writeAFS(data, fname)
}
\arguments{
\item{data}{AFS object to be saved}

\item{fname}{Save filename}
}
\description{
Write AFS
Write the AFS to a table
}
\examples{
filename<-system.file("extdata","sample.afs", package = "CCCA")
sampleAFS<-readAFS(filename)
writeAFS(sampleAFS,"sample.afs")
}

