% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeUDM.r
\name{writeUDM}
\alias{writeUDM}
\title{Write UDM}
\usage{
writeUDM(data, fname)
}
\arguments{
\item{data}{UDM object to be saved}

\item{fname}{Save filename}
}
\value{
Nothing
}
\description{
Write the unified data matrix to a file.
}
\examples{
filename<-system.file("extdata","sample.udm", package = "CCCA")
sampleUDM<-readUDM(filename)
writeUDM(sampleUDM,"sample.udm")
}

