% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.functions.r
\name{yaxis2.f}
\alias{yaxis2.f}
\title{Create a second y axis on the right side}
\usage{
yaxis2.f(x, y, ylabel = "2ndyaxis", cex = 0.75, ...)
}
\description{
Create a second y axis on the right side
}
\examples{

}
\keyword{function}
\keyword{helper}
