% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{addFasta}
\alias{addFasta}
\alias{addFasta.default}
\title{Add Fasta}
\usage{
addFasta(ccca, genome, width = 200, ...)

\method{addFasta}{default}(ccca, genome, width = 200, ...)
}
\arguments{
\item{ccca}{The object used to generate the fasta strings}

\item{genome}{A BS.genome object}

\item{width}{The width of fasta files, centered around the summit
of each peak in ccca}

\item{...}{Extension aguments}
}
\value{
ccca with fasta values
}
\description{
Sets the fasta value of ccca to the neuclotide values of the supplied
genome between start and end.
}
\examples{
#' @examples
## See ?makeCCCA
\dontrun{
library(BSgenome.Hsapiens.UCSC.hg19)
genome=BSgenome.Hsapiens.UCSC.hg19
addFasta(CCCA_prc,genome)
}
}
