% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{addFasta}
\alias{addFasta}
\alias{addFasta.default}
\title{AddReg Plot}
\usage{
addFasta(ccca, genome, width = 200, ...)

\method{addFasta}{default}()
}
\arguments{
\item{ccca}{The object used to generate the fasta strings}

\item{genome}{A BS.genome object}

\item{width}{The width of fasta files, centered around the summit of each peak in ccca}
}
\value{
ccca with fasta values
}
\description{
AddReg Plot
}
\examples{

}

