% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{addReg}
\alias{addReg}
\alias{addReg.default}
\title{Add Region}
\usage{
addReg(x, tag, logic, ...)

\method{addReg}{default}(x, tag, logic, ...)
}
\arguments{
\item{x}{The object with a $reg value to be added}

\item{tag}{The moniker of the new region to be added}

\item{logic}{A logical list of equal length to other regions in x$reg}

\item{...}{additional arguments}
}
\value{
An array is appended to <ccca>$reg
}
\description{
Add a region to the ccca object. These regions will be used as a mask to
identify seperate contexts.
}
\examples{
## See ?makeCCCA
}
