% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ffi-wrappers.r
\docType{package}
\name{CCCA}
\alias{CCCA}
\alias{CCCA-package}
\title{Compare multiple ChIP data sets across experimental conditions}
\description{
Compare multiple ChIP data sets across experimental conditions
}
\author{
"Alexander Griffith <griffitaj@gmail.com>"
}

