% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCCA.r
\name{generate}
\alias{generate}
\title{Generate}
\usage{
generate(peakFiles, alignedData, macsCutOff = 0, cl = NULL,
  categories = peakFiles, norm = qn)
}
\value{
a list containing <bed><over><heights><prc<normData><eigenVectors>>
}
\description{
Used to build the AFS and UDM.
}

