% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.data.ccf.functions.r
\name{Kobe.f}
\alias{Kobe.f}
\title{Make Kobe plots from simple data with overlay of another time series}
\usage{
Kobe.f(PB, E, Bref.multiplier = 1, col1 = "blue", col2 = "red", ...)
}
\arguments{
\item{PB}{the data and model fit coming from applying the PB model (PB.f)}

\item{E}{the time series to overlay and colour code Kobe years}

\item{Bref.multiplier}{a multiplier of the mean B in the reference period to adjust it to a specific reference point value}

\item{col1}{the colour at lowest value}

\item{col2}{the colour at the highest value}

\item{...}{par options for plot}
}
\description{
Make Kobe plots from simple data with overlay of another time series
}
\examples{
Kobe.f(PB=PB,E=PB$E)
}
\keyword{Bmsy,}
\keyword{Kobe,}
\keyword{environment}
\keyword{point,}
\keyword{reference}
