% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAFS.r
\name{makeAFS}
\alias{makeAFS}
\title{Make AFS}
\usage{
makeAFS(peakList, categories, format = "xls", pvalue = NULL, width = 700)
}
\arguments{
\item{peakList}{A list of strings refering to file locations containing Peak data}

\item{categories}{The monikers these peak files will be refered to ass}

\item{format}{The format to be imported, either bed or xls}

\item{pvalue}{The pvalue cut off to be applied, only valid with xls format}

\item{width}{The min width summits must be from one another in order to cluster}
}
\value{
The AFS generated from the input peak lists <chr><summit><catagory>
}
\description{
Take a list of peaks and create a unified set.
}
\examples{
peakList<-c("sample1.xls","sample2.xls","sample3.xls")
categories<-c("s1","s2","s3")
afs<-makeAFS(peakList,categories,pvalue=20)
}

