% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.r
\name{normalize}
\alias{normalize}
\alias{normalize.default}
\title{Normalize Plot}
\usage{
normalize(x, ...)

\method{normalize}{default}(x, ...)
}
\arguments{
\item{x}{Vector or matrix to be normalized}

\item{...}{additional arguments}
}
\value{
an array wher len = length(x)
}
\description{
Substract the mean and divide by the standard deviation.
}
\examples{
r<-runif(100,0,100)
normalize(r)
b=NULL
b$eigenVector<-matrix(r,10)
class(b)<-"PRC"
normalize(b)
}
