% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{normalize}
\alias{normalize}
\title{Normalize}
\usage{
normalize(x)
}
\arguments{
\item{x}{numeric input}
}
\value{
numeric of length x
}
\description{
privides unit normalization for numeric x
}
\examples{
x<-rnorm(100,10,2)
y<-normalize(x)
print(c(mean(y),sd(y)))
}

